requirements.txt
einops
ffmpeg
ffmpeg_python
GPUtil
moviepy
numpy
opencv_python
panda3d
pilgram
Pillow
PyMatting
PyQt6
pyqtgraph
qimage2ndarray
rawpy
requests
scikit_image
scipy
timm
torch
torchvision
tqdm

Flask
flask
numpy
google-generativeai
matplotlib
Pillow
altair<5
cv
opencv-python